feat: serde for cluster types and simpler impl of metric space
